requirements.txt
sentence-transformers>=2.2.0
PyPDF2>=3.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
psycopg2-binary>=2.9.0
requests>=2.31.0
hdbscan>=0.8.33
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
